requirements.txt
asgiref==3.7.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
icalendar==5.0.7
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
prettytable==3.8.0
psycopg2-binary==2.9.7
pyTelegramBotAPI==4.13.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
schedule==1.2.0
six==1.16.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
urllib3==2.0.4
Werkzeug==2.3.7
